% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{all_info_ns}
\alias{all_info_ns}
\title{Get all information on a namespace}
\usage{
all_info_ns(ns)
}
\arguments{
\item{ns}{A character representing the namespace to explore.}
}
\value{
A named list.
}
\description{
Get all information on a namespace
}
