% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/envir_deps.R
\name{deps}
\alias{deps}
\title{Get environment dependency object}
\usage{
deps(pkg, exported_only = FALSE)
}
\arguments{
\item{pkg}{A character vector of the environments to explore. Should be on the \link[base]{search} path. Cannot be \code{base}.}

\item{exported_only}{Whether to include non-exported functions.}
}
\value{
An object of class \code{pkgdepR}.
}
\description{
This function creates an object of class \code{pkgdepR}, representing the network of function calls between one or more packages.
}
\examples{
library(pkgdepR)
deps(pkg = "pkgdepR")
}
