% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{get_functions}
\alias{get_functions}
\title{Get functions from namespace}
\usage{
get_functions(ns)
}
\arguments{
\item{ns}{A character representing the namespace to explore.}
}
\value{
A character vector of function names.
}
\description{
Get functions from namespace
}
