% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3methods.R
\name{is.pkgdepR}
\alias{is.pkgdepR}
\title{pkgdepR object}
\usage{
is.pkgdepR(x)
}
\arguments{
\item{x}{Object to be tested.}
}
\value{
Boolean. \code{TRUE} when \code{x} is of class \code{pkgdepR}.
}
\description{
An internal generic function. Methods for \code{is.pkgdepR} should only return \code{TRUE} if the class is \code{pkgdepR}.
}
