% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{ls_namespace_info}
\alias{ls_namespace_info}
\title{Get namespace information}
\usage{
ls_namespace_info(ns, ...)
}
\arguments{
\item{ns}{A character representing the namespace to explore.}

\item{...}{Other arguments passed to \link[base]{ls}.}
}
\value{
A character vector of the names to search for a given namespace.
}
\description{
Get namespace information
}
