% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgdepR-package.R
\docType{package}
\name{pkgdepR-package}
\alias{pkgdepR}
\alias{pkgdepR-package}
\title{pkgdepR: Statically Determine Function Dependencies Between Packages}
\description{
An R package for (statically) determining and visualizing the intra-package and inter-package function dependencies. This may be useful for managing function dependencies across a code base of multiple R packages.
}
\details{
The only function you're likely to need from pkgdepR is \link[pkgdepR]{deps}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://pkgdepR.org/}
  \item Report bugs at \url{https://github.com/edpeyton/pkgdepR/issues}
}

}
\author{
\strong{Maintainer}: Ed Peyton \email{edppeyton@gmail.com} (\href{https://orcid.org/0000-0002-1427-0306}{ORCID})

}
\keyword{internal}
