% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3methods.R
\name{print.pkgdepR}
\alias{print.pkgdepR}
\title{Print pkgdepR object}
\usage{
\method{print}{pkgdepR}(x, ...)
}
\arguments{
\item{x}{An object of class \code{pkgdepR}.}

\item{...}{Redundant argument for consistency with method.}
}
\value{
No return value.
}
\description{
Print pkgdepR object
}
