% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3methods.R
\name{summary.pkgdepR}
\alias{summary.pkgdepR}
\title{Summarise pkgdepR object}
\usage{
\method{summary}{pkgdepR}(object, ...)
}
\arguments{
\item{object}{An object of class \code{pkgdepR}.}

\item{...}{Redundant argument for consistency with method.}
}
\value{
No return value.
}
\description{
Summarise pkgdepR object
}
