% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgdepR-package.R
\docType{package}
\name{pkgdepR-package}
\alias{pkgdepR}
\alias{pkgdepR-package}
\title{pkgdepR: Statically Determine Function Dependencies Between Packages}
\description{
Statically determine and visualize the function dependencies within and across packages. This may be useful for managing function dependencies across a code base of multiple R packages.
}
\details{
The only function you're likely to need from pkgdepR is \link[pkgdepR]{deps}, which returns an object of class \code{pkgdepR}.

An object of class \code{pkgdepR} is a list with three named objects:
\itemize{
 \item{\strong{\code{funs}}}{: a data frame describing the functions. Contains columns \code{id}, \code{label}, \code{package}, \code{exported}, \code{group}, and \code{name}.}
 \item{\strong{\code{links}}}{: a data frame containing the linkages between functions. Contains columns \code{from} and \code{to}.}
 \item{\strong{\code{pkg}}}{: a character vector containing the packages explored.}
}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://pkgdepR.org/}
  \item Report bugs at \url{https://github.com/edpeyton/pkgdepR/issues}
}

}
\author{
\strong{Maintainer}: Ed Peyton \email{edppeyton@gmail.com} (\href{https://orcid.org/0000-0002-1427-0306}{ORCID})

}
\keyword{internal}
